+++ packages/react-form-with-constraints/src/InputElement.ts
@@ -73,8 +73,10 @@ export class InputElement implements IHTMLInput {
 
   constructor(input: IHTMLInput | TextInput) {
     if (isHTMLInput(input)) {
+      // FIXME
       // eslint-disable-next-line no-param-reassign
       input = input as IHTMLInput;
+
       this.name = input.name;
       this.type = input.type;
       this.value = input.value;
@@ -90,8 +92,10 @@ export class InputElement implements IHTMLInput {
 
       this.validationMessage = input.validationMessage;
     } else {
+      // FIXME
       // eslint-disable-next-line no-param-reassign
       input = input as TextInput;
+
       this.name = input.props!.name;
       this.type = undefined as any;
       this.value = input.props!.value!; // Tested: TextInput props.value is always a string and never undefined (empty string instead)